Chore : 관련된 컴포넌트를 폴더에 정리, 파일 경로 수정
서로 연관된 컴포넌트를 일부 정리하고, 파일 경로를 수정하였습니다.
