+++ src/pages/Container/MypageContainer.jsx
@@ -5,16 +5,16 @@ import Profile from "components/main/Mypage/Profile/Profile"
 import Schedule from "components/main/Mypage/Schedule/Schedule"
 import Admin from "components/main/Mypage/Admin/Admin"
 import Greenbee from "components/main/Mypage/Greenbee/Greenbee"
-import GreenbeeInfo from "components/main/Mypage/Greenbee/GreenbeeInfo"
-import GreenbeeInfoEdit from "components/main/Mypage/Greenbee/GreenbeeInfoEdit"
+import GreenbeeInfo from "components/main/Mypage/Greenbee/Information/GreenbeeInfo"
+import GreenbeeInfoEdit from "components/main/Mypage/Greenbee/Information/GreenbeeInfoEdit"
 import MakeGreenbeeAccount from "components/main/Mypage/Greenbee/MakeGreenbeeAccount"
-import RequiredGreeningList from "components/main/Mypage/Greenbee/RequiredGreeningList"
-import RequiredGreeningRooftop from "components/main/Mypage/Greenbee/RequiredGreeningRooftop"
-import RequestToGreenBee from "components/main/RoofTop/RequestToGreenBee"
-import OtherGreenbeeInfo from "components/main/Mypage/Greenbee/OtherGreenbeeInfo"
+import RequiredGreeningList from "components/main/Mypage/Greenbee/RequiredGreening/RequiredGreeningList"
+import RequiredGreeningRooftop from "components/main/Mypage/Greenbee/RequiredGreening/RequiredGreeningRooftop"
+import RequestToGreenBee from "components/main/RoofTop/RequestGreenBee/RequestToGreenBee"
+import OtherGreenbeeInfo from "components/main/Mypage/Greenbee/Information/OtherGreenbeeInfo"
 import NotExistGreenbeeInfo from "components/main/Mypage/Greenbee/NotExistGreenbeeInfo"
 import Rooftop from "components/main/Mypage/RooftopOwner/Rooftop"
-import ApplyRoofTop from "components/main/RoofTop/ApplyRoofTop"
+import ApplyRoofTop from "components/main/RoofTop/ApplyRoofTop/ApplyRoofTop"
 import SuperviseRooftop from "components/main/RoofTop/Supervise/SuperviseRooftop"
 import SuperviseDetailRooftop from "components/main/RoofTop/Supervise/SuperviseDetailRooftop"
